Allow integration specs to be run via tag, rather than folder location.
